scripts/requirements.txt
requests
python-dotenv
tweepy
